+++ Assets/Scripts/Tasks/ReachToTargetTask.cs
@@ -108,7 +108,7 @@ public class ReachToTargetTask : BaseTask
         // Set up the target
 
         // Takes a target angle from the list and removes it
-        float targetAngle = ctrler.PollPseudorandomList("per_block_targetListToUse");
+        float targetAngle = (float) ctrler.PollPseudorandomList("per_block_targetListToUse");
         
         targets[2] = GameObject.Find("Target");
         targets[2].transform.rotation = Quaternion.Euler(
